# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zuu AI Video Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Videos ────────────────────────────────────────────────────────────────
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List character profiles
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharacterProfile"

    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a character profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterProfile"
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterProfile"

    put:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterProfile"

    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Providers ─────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/users/{id}/credits:
    post:
      operationId: adjustUserCredits
      tags: [admin]
      summary: Adjust user credits
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditAdjustInput"
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/jobs:
    get:
      operationId: listAdminJobs
      tags: [admin]
      summary: List generation jobs (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJobsResponse"
  /admin/providers:
    patch:
      operationId: updateProviderSettings
      tags: [admin]
      summary: Enable or disable a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderSettingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ─────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string

        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role, credits, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        credits:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]

    # ── Videos ───────────────────────────────────────────────────────────
...
    VideoGeneration:
      type: object
      required: [id, userId, prompt, style, aspectRatio, duration, quality, status, provider, isFavourite, creditsUsed, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        prompt:
          type: string
        enhancedPrompt:
          type: ["string", "null"]
        negativePrompt:
          type: ["string", "null"]
        style:
          type: string
        aspectRatio:
          type: string
        duration:
          type: integer
        quality:
          type: string
        cameraMovement:
          type: ["string", "null"]
        lighting:
          type: ["string", "null"]
        motionStrength:
          type: string
        seed:
          type: ["integer", "null"]
        provider:
          type: string
        providerJobId:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, queued, processing, completed, failed, cancelled]
        progressMessage:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        referenceImageUrl:
          type: ["string", "null"]
        characterProfileId:
          type: ["integer", "null"]
        isFavourite:
          type: boolean
        creditsUsed:
          type: integer
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CharacterProfile:
      type: object
      required: [id, userId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        age:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        faceDescription:
          type: ["string", "null"]
        hair:
          type: ["string", "null"]
        clothes:
          type: ["string", "null"]
        bodyType:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        animalType:
          type: ["string", "null"]
        cartoonStyle:
          type: ["string", "null"]
        referenceImageUrl:
          type: ["string", "null"]
        consistencyNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CharacterInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        faceDescription:
          type: ["string", "null"]
        hair:
          type: ["string", "null"]
        clothes:
          type: ["string", "null"]
        bodyType:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        animalType:
          type: ["string", "null"]
        cartoonStyle:
          type: ["string", "null"]
        referenceImageUrl:
          type: ["string", "null"]
        consistencyNotes:
          type: ["string", "null"]
    CharacterUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        faceDescription:
          type: ["string", "null"]
        hair:
          type: ["string", "null"]
        clothes:
          type: ["string", "null"]
        bodyType:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        animalType:
          type: ["string", "null"]
        cartoonStyle:
          type: ["string", "null"]
        referenceImageUrl:
          type: ["string", "null"]
        consistencyNotes:
          type: ["string", "null"]

    # ── Providers ────────────────────────────────────────────────────────
...
    AdminUsersResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUser:
      type: object
      required: [id, email, name, role, credits, totalGenerations, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        credits:
          type: integer
        totalGenerations:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminJobsResponse:
      type: object
      required: [items, total, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VideoGeneration"
        total:
          type: integer
        page:
          type: integer
    CreditAdjustInput:
      type: object
      required: [amount, description]
      properties:
        amount:
          type: integer
        description:
          type: string
    ProviderSettingUpdate:
      type: object
      required: [providerId, enabled]
      properties:
        providerId:
          type: string
        enabled:
          type: boolean